+++ requirements.txt
@@ -1,3 +1,4 @@
 Flask==1.1.2
 discord.py==1.3.3
 python-dotenv==0.13.0
+pytest==5.4.3
